# src/GuildSaber.Website/openapi/GuildSaber.Api.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "GuildSaber.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get all categories paginated.",
        "description": "Get all categories in the server.",
        "operationId": "GetCategoriesPaginated",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedListOfCategory"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{categoryId}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get a category.",
        "description": "Get a category in the server by its Id.",
        "operationId": "GetCategory",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.5.5",
                  "title": "Not Found",
                  "status": 404,
                  "instance": "GET categories/{categoryId}",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          }
        }
      }
    },
    "/guilds/{guildId}/categories": {
      "get": {
        "tags": [
          "Guilds.Categories"
        ],
        "summary": "Get all categories of a guild.",
        "description": "Get all categories of a guild by guild id.",
        "operationId": "GetCategories",
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Guilds.Categories"
        ],
        "summary": "Create a category for a guild.",
        "description": "Create a category for a guild by its Id.",
        "operationId": "CreateCategory",
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategory"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.5.1",
                  "title": "Bad Request",
                  "status": 400,
                  "detail": null,
                  "instance": "POST guilds/{guildId}/categories/",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
    "/guilds/{guildId}/categories/{categoryId}": {
      "put": {
        "tags": [
          "Guilds.Categories"
        ],
        "summary": "Update a category of a guild.",
        "description": "Update a category of a guild by its Id.",
        "operationId": "UpdateCategory",
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategory"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.5.1",
                  "title": "Bad Request",
                  "status": 400,
                  "detail": null,
                  "instance": "PUT guilds/{guildId}/categories/{categoryId}",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.5.5",
                  "title": "Not Found",
                  "status": 404,
                  "instance": "PUT guilds/{guildId}/categories/{categoryId}",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Guilds.Categories"
        ],
        "summary": "Delete a category of a guild.",
        "description": "Delete a category of a guild by its Id.",
        "operationId": "DeleteCategory",
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.5.5",
                  "title": "Not Found",
                  "status": 404,
                  "instance": "DELETE guilds/{guildId}/categories/{categoryId}",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/CategoryId"
          },
          "guildId": {
            "$ref": "#/components/schemas/GuildId"
          },
          "info": {
            "$ref": "#/components/schemas/CategoryInfo"
          }
        }
      },
      "CategoryId": {
        "type": "integer",
        "format": "int32"
      },
      "CategoryInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "CreateCategory": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "GuildId": {
        "type": "integer",
        "format": "int32"
      },
...
      "PagedListOfCategory": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "UpdateCategory": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "description": "GuildSaber session cookie. Browsers and cookie-enabled HTTP clients send it automatically.",
        "name": "__Host-GuildSaber.Session",
        "in": "cookie"
      },
      "X-GuildSaber-Request": {
        "type": "apiKey",
        "description": "Required for state-changing requests authenticated with the session cookie.",
        "name": "X-GuildSaber-Request",
        "in": "header"
      }
    }
  },
...
}